spell factory

+++ src/BadGuy.java
@@ -1,6 +1,6 @@
 public interface BadGuy {
     public void inflictDamage(int dmgInflicted);
-    public void reduceDamage(int dmgReduced);
-    public void reduceDefense(int defenseReduced);
-    public void reduceDodge(int dodgeReduced);
+    public void reduceDamage(double dmgReduced);
+    public void reduceDefense(double defenseReduced);
+    public void reduceDodge(double dodgeReduced);
 }
